refactor(PUPIL-421): move some downloadAndShareHelpers into SharedComponents
we're using these in both teacher and pupil views so there can be some
shared ownership of them.
