+++ src/components/PupilViews/PupilIntro/PupilIntro.view.test.tsx
@@ -11,10 +11,10 @@ import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
 import pupilLessonOverviewFixture from "@/node-lib/curriculum-api/fixtures/pupilLessonOverview.fixture";
 import { LessonEngineContext } from "@/components/PupilComponents/LessonEngineProvider";
 import { createLessonEngineContext } from "@/components/PupilComponents/LessonEngineProvider/LessonEngineProvider.test";
-import * as downloadLessonResources from "@/components/TeacherComponents/helpers/downloadAndShareHelpers/downloadLessonResources";
+import * as downloadLessonResources from "@/components/SharedComponents/helpers/downloadAndShareHelpers/downloadLessonResources";
 
 jest.mock(
-  "@/components/TeacherComponents/helpers/downloadAndShareHelpers/downloadLessonResources",
+  "@/components/SharedComponents/helpers/downloadAndShareHelpers/downloadLessonResources",
 );
 
 const curriculumData = pupilLessonOverviewFixture();
