+++ src/components/SharedComponents/helpers/downloadAndShareHelpers/createAndClickHiddenDownloadLink.test.tsx
