+++ src/components/SharedComponents/helpers/downloadAndShareHelpers/createDownloadResourcesLink.tsx
