+++ src/components/SharedComponents/helpers/downloadAndShareHelpers/downloadLessonResources.test.tsx
@@ -1,5 +1,5 @@
-import downloadLessonResources from "./downloadLessonResources";
 import createDownloadResourcesLink from "./createDownloadResourcesLink";
+import downloadLessonResources from "./downloadLessonResources";
 
 import type { ResourcesToDownloadArrayType } from "@/components/TeacherComponents/types/downloadAndShare.types";
