+++ src/components/SharedComponents/helpers/downloadAndShareHelpers/downloadLessonResources.tsx
@@ -1,5 +1,5 @@
-import createAndClickHiddenDownloadLink from "./createAndClickHiddenDownloadLink";
 import createDownloadResourcesLink from "./createDownloadResourcesLink";
+import createAndClickHiddenDownloadLink from "./createAndClickHiddenDownloadLink";
 
 import type { ResourcesToDownloadArrayType } from "@/components/TeacherComponents/types/downloadAndShare.types";
