+++ src/components/SharedComponents/helpers/downloadAndShareHelpers/getDownloadResourcesExistence.tsx
